rest_api/requirements.txt
fastapi
uvicorn
google-generativeai